requirements.txt
Django==1.8
uWSGI==2.0.10
dj-database-url==0.3.0
django-bootstrap3==5.1.1
django-getenv==1.3.1
psycopg2==2.6
python-dotenv==0.1.2
